Commit includes: Staker.approveAndRequestStake Facilitator.acceptStakeRequest
