+++ index.js
@@ -3,13 +3,17 @@
 const AbiBinProvider = require('./libs/AbiBinProvider');
 const EconomySetup = require('./libs/EconomySetup');
 const Contracts = require('./libs/Contracts');
-const StakeHelper = require('./libs/helpers/StakeHelper');
+const StakeHelper = require('./libs/helpers/stake/StakeHelper');
+const Staker = require('./libs/helpers/stake/gateway_composer/Staker');
+const Facilitator = require('./libs/helpers/stake/gateway_composer/Facilitator');
 
 module.exports = {
   AbiBinProvider: AbiBinProvider,
   EconomySetup: EconomySetup,
   Contracts: Contracts,
   Helpers: {
-    StakeHelper: StakeHelper
+    StakeHelper: StakeHelper,
+    Staker: Staker,
+    Facilitator: Facilitator
   }
 };
