+++ tests/helpers/BTHelper.js
@@ -100,14 +100,6 @@ describe('tests/helpers/BTHelper', function() {
       .liftRestriction([caOrganization, config.deployerAddress], config.facilitatorAddress)
       .then(validateReceipt);
   });
-
-  // TODO Is this needed?
-  // it('should lift restrictions from gateway and stakeVault addresses', function () {
-  //   let gateway = "0xe23338a00852250653Ba00563C4A38f559c4E5d5";
-  //   this.timeout(60000);
-  //   return helper
-  //     .setGateway(gateway, config.facilitatorAddress);
-  // });
 });
 
 // TODO Refactor to common method
