+++ tests/helpers/StakeHelper.js
@@ -11,7 +11,7 @@ const Setup = Package.EconomySetup,
   assert = chai.assert,
   config = require('../../tests/utils/configReader'),
   Web3WalletHelper = require('../../tests/utils/Web3WalletHelper'),
-  StakeHelper = require('../../libs/helpers/StakeHelper'),
+  StakeHelper = require('../../libs/helpers/stake/StakeHelper'),
   MockContractsDeployer = require('../utils/MockContractsDeployer'),
   abiBinProvider = MockContractsDeployer.abiBinProvider(),
   BTHelper = Package.EconomySetup.BrandedTokenHelper,
